pyproject.toml
[tool.poetry]
name = "pc2mqtt"
version = "0.1.0"
description = ""
authors = ["Maxim Mityutko"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
paho-mqtt = "1.6.1"

[tool.poetry.dev-dependencies]
pyinstaller = "6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
